% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_cycle_stats.R
\name{cv_cycle_stats}
\alias{cv_cycle_stats}
\title{Compute efficiencies for constant voltage cycling}
\usage{
cv_cycle_stats(inputdf, chargeid = "Charge", dischargeid = "Discharge",
  cycleid = "cycle", voltageid = "Vf", capacityid = "Q")
}
\arguments{
\item{inputdf}{Data frame containing all constant voltage cycles, preferably
imported via \code{import_all_cv_cycles()} or similar}

\item{chargeid}{Identifies charge cycles based on subsetting the input data
frame on the "type" column.}

\item{dischargeid}{Identifies discharge cycles based on subsetting the input
data frame.}

\item{cycleid}{The name of the column under which cycle number is stored in
the input data frame.}

\item{voltageid}{The name of the column under which the charging voltage is
stored in the input data frame.}

\item{capacityid}{The name of the column under which the charge capacity is
stored in the input data frame.}
}
\value{
Returns a data frame with columns for cycle, charge and discharge
  capacities, current efficiency, charging and discharging voltages,
  charging and discharging energies, and energy efficiency.
}
\description{
This function takes as input a data set of constant voltage cycles and
returns a data frame listing cycle number, capacity, current efficiency,
energy efficiency, and charging and discharging voltages.
}
\details{
At time of writing, Gamry Framework doesn't allow constant voltage cycling
with its cyclic charge/discharge package, meaning that a single .DTA file
with the capacities, efficiencies, etc for constant voltage cycling is never
generated. This function collects that data from constant voltage cycles.
The corollary for constant current cycles has not been written; instead
the \code{import_cycling_stats()} function can be used.
}
\author{
Mike Gerhardt
}
\seealso{
\code{\link{import_all_cv_cycles}}, \code{\link{import_cycling_stats}}
}

