% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eChem-package.R
\docType{package}
\name{eChem-package}
\alias{eChem-package}
\title{Simulations for Electrochemistry Experiments}
\description{
A collection of functions for simulating four different types of
electrochemistry experiments: two potential step experiments
(chronoamperometry and chronocoulometry) and two potential scan experiments
(linear-sweep voltammetry and cyclic voltammetry). Each simuluation allows
for either an oxidation reaction or for a reduction reaction, and allows for
a single following or a single preceeding chemical reaction. Also included
are a variety of functions for displaying the result of a simulation in
either a static or animated format.
}
\author{
David T. Harvey

Maintainer: David T. Harvey \email{harvey@depauw.edu}
}
