% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_graph.R
\name{export_graph}
\alias{export_graph}
\title{Save a ggplot graph}
\usage{
export_graph(fname, ggname, width = 600, height = 360)
}
\arguments{
\item{fname}{Filename to which export should occur. If the file doesn't yet
exist, it is created.}

\item{ggname}{Name of the gg object to export.}
}
\description{
Exports a ggplot graph with a set width and height
}
\author{
Mike Gerhardt
}

