% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_absorption.R
\name{import_absorption}
\alias{import_absorption}
\title{Import absorption spectrum.}
\usage{
import_absorption(filename = file.choose(),
  data_tag = ">>>>>Begin Spectral Data<<<<<", columns = c("Wavelength",
  "Absorption"))
}
\arguments{
\item{filename}{The file to import.}

\item{data_tag}{Identifies where in the file the data begins, defaults to
the right option for the Ocean Optics Flame spectrometer in our lab.}

\item{columns}{What to name the columns in the data frame.}
}
\value{
Returns a data frame with column names given by the \code{columns}
  input parameter.
}
\description{
This function imports an absorption spectrum collected by the Ocean Optics
Flame spectrometer in our lab. It just needs to be told where the data
begins by the data_tag parameter.
}
\author{
Mike Gerhardt
}

