% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_cc_cycles.R, R/import_all_cv_cycles.R
\name{import_all_cc_cycles}
\alias{import_all_cc_cycles}
\alias{import_all_cv_cycles}
\title{Import each cycle in a folder.}
\usage{
import_all_cc_cycles(filepath = "mrgPWRCAPACITY/CHARGE_DISCHARGE/",
  allcyclesid = "CHARGE", dischargeid = "DISCHARGE", ...)

import_all_cv_cycles(filepath, allcyclesid = "charge",
  dischargeid = "discharge", ...)
}
\arguments{
\item{filepath}{Path where the charge/discharge cycle files are located}

\item{allcyclesid}{A character string which is compared to each filename
in \code{filepath}. Any file having this string is considered a cycle.}

\item{dischargeid}{A character string which is compared to each cycle
found by \code{allcyclesid}. All cycles having this pattern are marked
discharge cycles, and the remaining cycles are assumed to be charge
cycles.}

\item{...}{arguments to pass to each import_cycle function.}
}
\value{
Returns a data frame containing each cycle in a folder, with a
  cycle number tag allowing you to subset the data by cycle if you wish.
}
\description{
These functions imports all the constant current or constant voltage cycling
data in a given folder. The functions rely on differences in filename to
distinguish charge curves from discharge curves.
}
\details{
Each of these functions calls the appropriate import_cc_cycle or 
import_cv_cycle fuction on each file matching the descriptions you give in
the function call.
}
\section{Functions}{
\itemize{
\item \code{import_all_cv_cycles}: Import all constant voltage cycling curves
(current vs time)
}}
\author{
Mike Gerhardt
}
\seealso{
\code{\link{import_cc_cycle}}, \code{\link{import_cv_cycle}}
}

