% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_all_cyclic_voltammetry.R
\name{import_all_cyclic_voltammetry}
\alias{import_all_cyclic_voltammetry}
\title{Imports all cyclic voltammetry data}
\usage{
import_all_cyclic_voltammetry(filepath, identifier = ".DTA", ...)
}
\arguments{
\item{filepath}{The directory in which the files are located.}

\item{identifier}{A character string. All file names in \code{filepath} are 
searched for this value, and any file names with a match are imported.}

\item{...}{More parameters to pass to import_cv.}
}
\value{
Returns a list. Each list element contains the full path and file
  name of the imported file and a data frame with that file's data.
}
\description{
Imports data from any cyclic voltammetry curves collected with the Gamry
potentiostat and outputs them in a list of data frames.
}
\details{
This function finds all files marked with \code{identifier}, assumes they
are cyclic voltammograms, and calls \code{import_cv} on each. It puts
out a list, with each list element containing the full filename from which
the data was imported and the data itself in a data frame.
This function should NOT be used with cyclic voltammograms performed on
a full cell to obtain polarization curves. Use \code{import_cell_cv} 
instead.
}
\author{
Mike Gerhardt
}
\seealso{
\code{\link{import_cv}}
}

