% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cc_cycle.R
\name{import_cc_cycle}
\alias{import_cc_cycle}
\title{Imports a single constant-current cycle.}
\usage{
import_cc_cycle(inputfile = file.choose(), cols = c("T", "Vf", "Im"),
  getrealtime = TRUE, ...)
}
\arguments{
\item{inputfile}{File to import.}

\item{cols}{The names of the columns to import. These must match the names
given by Gamry Framework to the data in the file.}

\item{getrealtime}{If TRUE, searches for the start time offset (see details)}
}
\value{
Returns a data frame containing the time, voltage, current, and
charge sampled during the cycle.
}
\description{
This function takes a single charge or discharge cycle from a cyclic
charge/discharge loop run from Gamry Framework and outputs the time,
voltage, and current into a data frame.
}
\details{
This function calls \code{importGamry} and strips the output down to
time, voltage, and current columns. It then multiplies time by current
to get charge passed and adds that as a fourth column. If \code{getrealtime}
is TRUE then the function will find the last line of the Gamry data file,
where the "start time offset" is listed - this is how much time has passed
since the first cycle in the sequence.
}
\author{
Mike Gerhardt
}
\seealso{
\code{import_all_cc_cycles}, \code{import_cycling_stats},
  \code{cv_cycle_stats}
}

