% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cell_cv.R, R/polcurve_averager.R
\name{import_cell_cv}
\alias{import_cell_cv}
\alias{polcurve_averager}
\alias{polcurve_smoother}
\title{Imports full cell cyclic voltammogram}
\usage{
import_cell_cv(fname = file.choose(), area = 5, rpm = 50, cols = c("T",
  "Vf", "Im"), movingavg = TRUE, traceavg = TRUE, ...)

polcurve_averager(inputdf, voltageid = "Vf", currentid = "scurrent")

polcurve_smoother(inputdf, rpm)
}
\arguments{
\item{fname}{File to import.}

\item{area}{Geometric area of the cell electrodes, for correction from
current/power to current/power density.}

\item{rpm}{Rotation rate of the pumps, in rotations per minute. This is
used by the moving average filter to average the collected data over
one pump pulse.}

\item{cols}{Data columns to import. Must match the names given by Gamry
Framework software.}

\item{movingavg}{If TRUE (default is TRUE), the function will apply a
moving average to the measured current to smooth out oscillations in
the data from a pulsing pump.}

\item{traceavg}{If TRUE (default is TRUE), the function will attempt to
split the polarization curve into two traces, a top and bottom trace, and
average the two. This is useful if the cyclic voltammogram is changing
the state of charge of the cell.}
}
\value{
Returns a data frame with current, voltage, and power columns. The
   filename of each file is also reported in the state of charge ("soc")
   column for later analysis.
}
\description{
This function imports a cyclic voltammogram performed on a full cell for the
 purpose of measuring a polarization curve.
}
\details{
This function, by default, performs two filters on the data to smooth it.
 First, a moving average filter is applied to dampen pulses from non-steady
 flow pumps such as diaphragm pumps. The moving average is applied over one
 pump cycle, which is why the function takes pump RPM as an input. The 
 second filter splits the polarization curve in half and averages the top
 and bottom halves, to adjust for state of charge changes while running the
 cyclic voltammogram.
}
\section{Functions}{
\itemize{
\item \code{polcurve_averager}: Function which averages the top and bottom traces
of a polarization curve.

\item \code{polcurve_smoother}: Function which smooths out the pulses in the
polarization curve due to diaphragm pumps
}}
\author{
Mike Gerhardt
}

