% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cv_cycle.R
\name{import_cv_cycle}
\alias{import_cv_cycle}
\title{Imports a single constant-voltage cycle.}
\usage{
import_cv_cycle(inputfile = file.choose(), cols = c("T", "Q", "Vf", "Im"),
  ...)
}
\arguments{
\item{inputfile}{File to import.}

\item{cols}{The names of the columns to import. These must match the names
given by Gamry Framework to the data in the file.}
}
\value{
Returns a data frame containing the time, voltage, current, and
charge sampled during the cycle.
}
\description{
This function takes a single charge or discharge cycle from a cyclic
charge/discharge loop run from Gamry Framework and outputs the time,
voltage, and current into a data frame.
}
\details{
This function calls \code{importGamry} and strips the output down to
time, charge, voltage, and current columns. Unlike \code{import_cc_cycle}
this function can't take a sequence of cycles and output during each
cycle the time elapsed from the start of the first cycle.
}
\author{
Mike Gerhardt
}
\seealso{
\code{\link{import_all_cc_cycles}}, \code{\link{import_cycling_stats}},
  \code{\link{cv_cycle_stats}}
}

