% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ocp.R
\name{import_ocp}
\alias{getocp}
\alias{import_ocp}
\title{Import open circuit potential vs state of charge}
\usage{
import_ocp(filepattern = "OCP_before_CV", state.of.charge = seq(from = 10,
  to = 100, by = 10), filepath = ".")

getocp(filename)
}
\arguments{
\item{filepattern}{A unique tag which appears in the filename of each data
file you wish to import}

\item{state.of.charge}{A sequence of numbers which should be of the same
length as the number of files you wish to import. Each number in the 
sequence should correspond to the state of charge at which you made each
open circuit potential measurement.}

\item{filepath}{Path where the open circuit potential vs state of charge data
is stored. Defaults to working directory.}
}
\value{
Returns a data frame containing open circuit potential and state of
  charge.
}
\description{
This function, given a set of open circuit potential measurement files from
Gamry Framework, collects the last voltage data point from each file, and
puts them into a data frame to make plotting open circuit potential versus
state of charge straightforward
}
\details{
The function requires all open circuit potential files to have a unique
character string \code{filepattern} in their filenames. The function also
currently requires the user to specify the state of charge at which each
data file was collected. The function looks in the working directory for any
files with the tag \code{filepattern} and then uses \code{mixedsort} to
intelligently sort the files (so that, for example, #10 does not come
between #1 and #2).
}
\section{Functions}{
\itemize{
\item \code{getocp}: This function returns the last recorded voltage from
an open circuit potential data file.
}}
\author{
Mike Gerhardt
}
\seealso{
\code{\link[gtools]{mixedsort}}
}

