% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGamry.R
\name{importGamry}
\alias{importGamry}
\title{Imports a Gamry data file.}
\usage{
importGamry(inputfile = file.choose(), ncols = 0)
}
\arguments{
\item{inputfile}{File to be imported.}

\item{ncols}{Number of expected data columns. The default of 0 has the
function make an educated guess about how many columns there are.}
}
\value{
A data frame with all the data in the .dta file.
}
\description{
\code{importGamry} imports a Gamry .dta file and returns a data frame with 
each column corresponding to a column in the data file.
}
\details{
This function needs to know which lines in the .dta file to import. The 
default behavior is to count each line in the file and assume that the
number that appears the most times is where the data is stored in the file.
Sometimes this fails because Gamry adds extensive header documentation to
their files. The ncols parameter can be used to manually specify which lines
to import based on how many columns each line has.
}
\author{
Mike Gerhardt
}

