% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_capacities.R
\name{plot_capacities}
\alias{plot_capacities}
\title{Plot charge and discharge capacity versus cycle number}
\usage{
plot_capacities(inputdf, xdata = "Cycle", ydata = "Charge")
}
\arguments{
\item{inputdf}{Data frame with the data to plot. Should be a data frame made
with \code{import_cycling_stats}}

\item{xdata}{Name of the column under which the cycle numbers are stored.}

\item{ydata}{Name of the column under which the charge/discharge capacities
are stored.}
}
\value{
Returns a ggplot object plotting charge/discharge capacity versus
  cycle number.
}
\description{
This function takes a data frame with charge and discharge capacity data and
makes a plot of these capacities versus cycle number. Useful for qualitative
investigations of capacity fade.
}
\details{
This function should only be used on data frames generated with
\code{import_cycling_stats} as it will probably fail otherwise. It assumes
that the Type column exists in the input data frame and colors the plot by
Type. The function will return a ggplot, which can be further edited through
any of the ggplot methods.
}
\author{
Mike Gerhardt
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{import_cycling_stats}}
}

