% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycles.R
\name{plot_cycles}
\alias{plot_cycles}
\title{Plot a set of constant-current charge-discharge cycles
 
 This function takes a set of voltage vs. time curves from a cyclic
 charge-discharge experiment and plots them.
 
 This function is designed to be used on data frames generated with
 \code{import_all_cc_cycles} and may not work otherwise. The function will
 return a ggplot object which can be modified with ggplot methods.
 
 @param inputdf Data frame, typically generated with \code{import_all_cc_cycles}
 
 @author Mike Gerhardt
 @return Returns a ggplot object containing all charge/discharge cycles
   specified by \code{cyclelist}
 
 @export
 @seealso \code{\link{import_all_cc_cycles}}}
\usage{
plot_cycles(inputdf, cyclelist = "all")
}
\description{
Plot a set of constant-current charge-discharge cycles
 
 This function takes a set of voltage vs. time curves from a cyclic
 charge-discharge experiment and plots them.
 
 This function is designed to be used on data frames generated with
 \code{import_all_cc_cycles} and may not work otherwise. The function will
 return a ggplot object which can be modified with ggplot methods.
 
 @param inputdf Data frame, typically generated with \code{import_all_cc_cycles}
 
 @author Mike Gerhardt
 @return Returns a ggplot object containing all charge/discharge cycles
   specified by \code{cyclelist}
 
 @export
 @seealso \code{\link{import_all_cc_cycles}}
}

