% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_efficiencies.R
\name{plot_efficiencies}
\alias{plot_efficiencies}
\title{Plot current and energy efficiency versus cycle number}
\usage{
plot_efficiencies(inputdf, energyeffy = TRUE)
}
\arguments{
\item{inputdf}{Data frame with the data to plot. Should be a data frame made
with \code{import_cycling_stats}}

\item{energyeffy}{Defaults to TRUE. If TRUE, plot the energy efficiency as
well as the current efficiency.}
}
\value{
Returns a ggplot object plotting current/energy efficiency versus
  cycle number.
}
\description{
This function takes a data frame with current and energy efficiency data and
makes a plot of these efficiencies versus cycle number.
}
\details{
This function is designed to be used on data frames generated with
\code{import_cycling_stats} and assumes that the Cycle and ceffy columns exist.
The function will return a ggplot, which can be further edited through
any of the ggplot methods.
}
\author{
Mike Gerhardt
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{import_cycling_stats}}
}

