\name{echem-package}
\alias{echem-package}
\alias{echem}
\docType{package}
\title{
Importing and manipulating data from Gamry potentiostats
}
\description{
Import .DTA files output from Gamry potentiostats during cyclic voltammetry, charge/discharge cycling, etc
}
\details{
\tabular{ll}{
Package: \tab echem\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-17\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mike Gerhardt

Maintainer: Mike Gerhardt <gerhardt.mike@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
importGamry()
}
