% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cc_cycle.R
\name{import_cc_cycle}
\alias{import_cc_cycle}
\title{Imports a single constant-current cycle.}
\usage{
import_cc_cycle(inputfile = file.choose(), cols = c("T", "Vf", "Im"),
  getrealtime = TRUE, ...)
}
\description{
Imports a single constant-current cycle.
}

